+++ AnimalShop/AnimalShop/Pterodactyl.cs
@@ -6,7 +6,7 @@
         public float Wingspan { get; private set; }
         public float BeakLength { get; private set; }
 
-        public Pterodactyl(string featherColor, float wingspan, float beakLength, int age, double weight, float cost) : base(age, weight, cost)
+        public Pterodactyl(string featherColor, float wingspan, float beakLength, int age, double weight, decimal cost) : base(age, weight, cost)
         {
             FeatherColor = featherColor;
             Wingspan = wingspan;
